Modifications necessary to make sure all document types are displayed when a user clicks on the filter icon.  Also modified the by grants page so that when a user selects the magnifying glass icon that only documents on grants matching the the filters selected get displayed.

+++ source-aspnet/eGrants/Models/doclayer.cs
@@ -81,7 +81,7 @@ namespace eGrants.Models
         /// <summary>
         ///     Gets or sets the category_id.
         /// </summary>
-        public int? category_id { get; set; }
+        public short? category_id { get; set; }
 
         /// <summary>
         ///     Gets or sets the category_name.
@@ -126,7 +126,7 @@ namespace eGrants.Models
         /// <summary>
         ///     Gets or sets the page_count.
         /// </summary>
-        public int? page_count { get; set; }
+        public short? page_count { get; set; }
 
         /// <summary>
         ///     Gets or sets the fsr_count.
